@@ -4,7 +4,7 @@
     "type": "library",
     "license": "MIT",
     "require": {
-        "php": ">=7.2.0"
+        "php": ">=7.4"
     },
     "require-dev": {
         "ext-pdo": "*",
@@ -20,6 +20,6 @@
     },
     "scripts": {
         "test": "phpunit tests",
-        "check": "psalm --no-cache"
+        "check": "psalm"
     }
 }
